+++ src/client/components/features/JeRecruteAfprPoei/RecrutementCandidatFranceTravail/RecrutementCandidatFranceTravail.test.tsx
@@ -32,7 +32,7 @@ describe('<RecrutementCandidatFranceTravail/>', () => {
 
 		render(<RecrutementCandidatFranceTravail/>);
 
-		expect(screen.getByRole('link', { name: 'M‘engager à recruter' })).toHaveAttribute('href', LINK_M_ENGAGER_A_RECRUTER);
-		expect(screen.getByRole('link', { name: 'Compléter une demande' })).toHaveAttribute('href', LINK_COMPLETER_UNE_DEMANDE);
+		expect(screen.getByRole('link', { name: 'M‘engager à recruter - nouvelle fenêtre' })).toHaveAttribute('href', LINK_M_ENGAGER_A_RECRUTER);
+		expect(screen.getByRole('link', { name: 'Compléter une demande - nouvelle fenêtre' })).toHaveAttribute('href', LINK_COMPLETER_UNE_DEMANDE);
 	});
 });