+++ src/client/components/features/ServiceCard/MétiersDuSoinPartner.tsx
@@ -7,6 +7,7 @@ export function MétierDuSoinPartner() {
 		<ServiceCard
 			linkLabel="En savoir plus"
 			logo="/images/logos/métiers-du-soin.svg"
+			logoAlt="Les métiers du soin et de l’accompagnement recrutent"
 			link="https://solidarites-sante.gouv.fr/metiers-et-concours/metiers-soin-et-accompagnement/metiersdusoin"
 			title="Renseignez-vous sur les métiers du soin"
 			titleAs={'h3'}