+++ src/pages/mentorat/index.page.test.tsx
@@ -69,7 +69,7 @@ describe('MentoratPage', () => {
 				<MentoratPage />
 			</DependenciesProvider>,
 		);
-		const linkAsButton = screen.getByRole('link', { name: 'Trouver mon mentor' });
+		const linkAsButton = screen.getByRole('link', { name: 'Trouver mon mentor - nouvelle fenêtre' });
 
 		expect(linkAsButton).toBeVisible();
 		expect(linkAsButton).toHaveAttribute('href', 'https://www.1jeune1mentor.fr/formulaire?1jeune1solution');